+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Unreleased
 
+* **Breaking** `FileDialogFilters::push_filter` now accepts any `IntoIterator<Item=str>`.
+
+* Better runtime info about supported image formats.
+    - **Breaking** Removed `IMAGES.available_encoders/decoders`.
+    - Added `IMAGES.available_formats` with `ImageFormat`.
+    - Format info includes capabilities (encode, entries), display name and a list of file extensions and media types.
+    - **Breaking** Unified view-process API capabilities info into a single struct.
+
+* Fix `ImageVar` updates not propagating in some complex bindings.
+
 * Fix `cargo zng l10n --clean-template` removing custom template localization files.
 * Fix `cargo zng l10n` not including entries of local dependencies targeting the same localization file.